Fix LGTM warnings

+++ src/OVAL/probes/unix/linux/systemdshared.h
@@ -261,10 +261,10 @@ static char *dbus_value_to_string(DBusMessageIter *iter)
 
 #ifdef DBUS_HAVE_INT64
 			case DBUS_TYPE_INT64:
-				return oscap_sprintf("%lli", value.i32);
+				return oscap_sprintf("%lli", value.i64);
 
 			case DBUS_TYPE_UINT64:
-				return oscap_sprintf("%llu", value.u32);
+				return oscap_sprintf("%llu", value.u64);
 #endif
 
 			case DBUS_TYPE_DOUBLE:
